Вынес компоненты в отдельные файлы
ProductCategoryPage и ProductSubcategoryPage
