+++ src/routes.jsx
@@ -6,48 +6,11 @@ import {
     ContactsPage,
     FooterComponent,
     PageContainer,
-    GridView,
+    ProductCategoryPage,
+    ProductSubcategoryPage
 } from './components'
 
 import {categoriesData} from './data'
-import {gql, useQuery} from "@apollo/client";
-
-
-const ProductCategoryPage = ({category}) => {
-    return (
-        <div>
-            <h3 style={{textAlign: "center"}}>{category.name}</h3>
-            <GridView cardItems={category.subcategories}/>
-        </div>
-    )
-}
-
-const ProductSubcategoryPage = ({subcategory}) => {
-    const ProductsQuery = gql`
-        query ProductsQuery($categoryName: String!) {
-             categoryProducts(categoryName: $categoryName) {
-                description
-                id
-                name
-                images {
-                  url           
-                }
-            }
-        }
-    `
-    const {loading, error, data} = useQuery(ProductsQuery, {
-        variables: {categoryName: subcategory.name},
-    })
-
-    if (loading) return <p>Loading...</p>
-    if (error) return <p>Error</p>
-    return (
-        <div>
-            <h3 style={{"textAlign": "center"}}>{subcategory.name}</h3>
-            <GridView apiPatterns cardItems={data.categoryProducts}/>
-        </div>
-    )
-}
 
 export const useRoutes = () => {
     return (
